+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project won't have versionining per se. But it's format is based in the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2022-07-31
+
+### Added
+
+- Created the postman to markdown utility
+
 ## 2022-07-29
 
 ### Added